# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hệ Thống Tự Động Sinh Test Case API (CDIO-4)",
    "description": "Nền tảng tự động bóc tách đặc tả yêu cầu, giải ràng buộc và sinh bộ kiểm thử phần mềm đa phương pháp (BVA, Phân vùng tương đương, Pairwise, Z3 Solver). Quản lý chu kỳ Scrum, phân công nhiệm vụ thành viên và xuất báo cáo kiểm thử nghiệm thu chuyên nghiệp.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/requirements/project/{project_id}": {
      "get": {
        "tags": [
          "Chức năng & Yêu cầu"
        ],
        "summary": "Danh sách chức năng con trong dự án",
        "description": "Lấy danh sách tất cả các chức năng con (User Stories / Requirements) trong dự án kèm chỉ số kiểm thử.",
        "operationId": "get_project_requirements_api_requirements_project__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequirementOut"
                  },
                  "title": "Response Get Project Requirements Api Requirements Project  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements": {
      "post": {
        "tags": [
          "Chức năng & Yêu cầu"
        ],
        "summary": "Tạo chức năng con / User Story mới",
        "description": "Tạo một chức năng mới trong dự án kèm đặc tả yêu cầu, người thực hiện, thời hạn và độ ưu tiên.",
        "operationId": "create_requirement_api_requirements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{requirement_id}": {
      "put": {
        "tags": [
          "Chức năng & Yêu cầu"
        ],
        "summary": "Cập nhật chức năng, phân công và trạng thái",
        "description": "Chỉnh sửa tiêu đề, đặc tả yêu cầu, phân công người kiểm thử (Assignee), thời hạn (Due Date) hoặc cập nhật tiến độ (planning / in_testing / completed).",
        "operationId": "update_requirement_api_requirements__requirement_id__put",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chức năng & Yêu cầu"
        ],
        "summary": "Xóa chức năng con",
        "description": "Xóa vĩnh viễn chức năng con cùng toàn bộ tham số và các bộ ca kiểm thử liên quan.",
        "operationId": "delete_requirement_api_requirements__requirement_id__delete",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/generate": {
      "post": {
        "tags": [
          "Sinh Ca Kiểm Thử"
        ],
        "summary": "Sinh bộ ca kiểm thử tự động theo kỹ thuật",
        "description": "Nhận danh sách tham số và lựa chọn kỹ thuật kiểm thử (BVA, Phân vùng tương đương, Pairwise, Z3 Solver) để sinh bộ test case hoàn chỉnh.",
        "operationId": "generate_test_cases_api_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/testcases/{tc_id}": {
      "patch": {
        "tags": [
          "Sinh Ca Kiểm Thử"
        ],
        "summary": "Cập nhật kết quả kiểm thử thực tế (Pass/Fail)",
        "description": "Cập nhật trạng thái thực thi (Pass, Fail, Blocked, Untested) và kết quả thực tế (Actual Result) của ca kiểm thử.",
        "operationId": "update_test_case_api_testcases__tc_id__patch",
        "parameters": [
          {
            "name": "tc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suites/{suite_id}": {
      "get": {
        "tags": [
          "Sinh Ca Kiểm Thử"
        ],
        "summary": "Chi tiết bộ Test Suite và danh sách ca kiểm thử",
        "description": "Lấy chi tiết bộ kiểm thử (TestSuite) kèm độ bao phủ Pairwise và toàn bộ các ca kiểm thử bên trong.",
        "operationId": "get_suite_details_api_suites__suite_id__get",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "properties": {
          "requirement_id": {
            "type": "integer",
            "title": "Requirement Id"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ParameterBase"
            },
            "type": "array",
            "title": "Parameters"
          },
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques",
            "default": [
              "bva",
              "ep",
              "pairwise"
            ]
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "requirement_id",
          "parameters"
        ],
        "title": "GenerateRequest",
        "description": "Client gui danh sach tham so (da chinh sua) de sinh test cases."
      },
      "GenerateResponse": {
        "properties": {
          "test_suite_id": {
            "type": "integer",
            "title": "Test Suite Id"
          },
          "requirement_id": {
            "type": "integer",
            "title": "Requirement Id"
          },
          "total_cases": {
            "type": "integer",
            "title": "Total Cases"
          },
          "coverage_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage Info",
            "default": {}
          },
          "test_cases": {
            "items": {
              "$ref": "#/components/schemas/TestCaseOut"
            },
            "type": "array",
            "title": "Test Cases"
          }
        },
        "type": "object",
        "required": [
          "test_suite_id",
          "requirement_id",
          "total_cases",
          "test_cases"
        ],
        "title": "GenerateResponse",
        "description": "Ket qua sinh test cases."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ParameterBase": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          },
          "min_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Length"
          },
          "max_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values"
          },
          "regex_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex Pattern"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": true
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "data_type"
        ],
        "title": "ParameterBase"
      },
...
      "RequirementCreate": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "format_type": {
            "type": "string",
            "title": "Format Type",
            "default": "free_text"
          },
          "sprint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Id"
          },
          "assignee_name": {
            "type": "string",
            "title": "Assignee Name",
            "default": ""
          },
          "due_date": {
            "type": "string",
            "title": "Due Date",
            "default": ""
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "Medium"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "planning"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title",
          "raw_text"
        ],
        "title": "RequirementCreate"
      },
      "RequirementOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "sprint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "format_type": {
            "type": "string",
            "title": "Format Type"
          },
          "assignee_name": {
            "type": "string",
            "title": "Assignee Name",
            "default": ""
          },
          "due_date": {
            "type": "string",
            "title": "Due Date",
            "default": ""
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "Medium"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "planning"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "parameters_count": {
            "type": "integer",
            "title": "Parameters Count",
            "default": 0
          },
          "test_suites_count": {
            "type": "integer",
            "title": "Test Suites Count",
            "default": 0
          },
          "total_cases": {
            "type": "integer",
            "title": "Total Cases",
            "default": 0
          },
          "pass_count": {
            "type": "integer",
            "title": "Pass Count",
            "default": 0
          },
          "fail_count": {
            "type": "integer",
            "title": "Fail Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "raw_text",
          "format_type"
        ],
        "title": "RequirementOut"
      },
      "RequirementUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "format_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Type"
          },
          "sprint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Id"
          },
          "assignee_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Name"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "RequirementUpdate"
      },
...
      "TestCaseOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "preconditions": {
            "type": "string",
            "title": "Preconditions",
            "default": ""
          },
          "test_steps": {
            "type": "string",
            "title": "Test Steps",
            "default": ""
          },
          "test_type": {
            "type": "string",
            "title": "Test Type"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data"
          },
          "expected_result": {
            "type": "string",
            "title": "Expected Result"
          },
          "actual_result": {
            "type": "string",
            "title": "Actual Result",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Untested"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "Medium"
          },
          "technique_source": {
            "type": "string",
            "title": "Technique Source",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "scenario",
          "test_type",
          "input_data",
          "expected_result"
        ],
        "title": "TestCaseOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}